requirements.txt
# === Core Data Handling & Numerics ===
pandas>=2.0.0
numpy>=1.24.0

# === Web App & UI ===
streamlit>=1.28.0
plotly>=5.17.0
streamlit-option-menu>=0.3.6

# === Core Data Ingestion (APIs & Scraping) ===
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# === Financial Data APIs ===
yfinance>=0.2.28
ccxt>=4.0.0              # Crypto exchange data & arbitrage scanning
alpha-vantage            # Fundamental data (500 req/day free)
fredapi                  # Federal Reserve economic data (unlimited free)
eia-python               # Energy Information Administration data
finnhub-python           # Insider trades, news, financial data

# === Social Media & News APIs ===
praw>=7.7.0              # Reddit API
newsapi-python           # News sentiment (100 req/day free)
textblob>=0.19.0         # Sentiment analysis
nltk>=3.9                # Natural language processing
vaderSentiment>=3.3.2    # Financial sentiment analysis

# === Machine Learning & Forecasting ===
scikit-learn>=1.3.0      # ML models, metrics, backtesting
scipy>=1.11.0            # Statistical tests
statsmodels              # Cointegration testing, time series analysis
transformers             # Advanced NLP (HuggingFace)
torch                    # PyTorch backend for transformers
prophet                  # Time-series forecasting (Meta's Prophet)
anthropic                # Claude LLM API integration

# === Financial & Risk Analysis ===
ta>=0.11.0               # Technical Analysis library
py_vollib>=1.0.1         # Options Greeks calculations
# Portfolio Analysis
quantstats               # Backtest metrics and portfolio analysis (replaces pyfolio)

# === Global AI Analysis (Phase 3) ===
aiohttp>=3.9.0           # Async HTTP client for parallel API calls
openai>=1.0.0            # GPT-4 API client
google-generativeai      # Gemini Pro API client

# === Storage & Caching ===
sqlalchemy>=2.0.0        # Database ORM
pyarrow>=14.0.0          # Parquet file storage

# === Utils ===
python-dotenv>=1.0.0     # Environment variable management
